requirements.txt
fastapi
uvicorn[standard]
pydantic
sentence-transformers
faiss-cpu
pypdf
openai
streamlit
python-dotenv
